+++ nodetools/configuration/constants.py
@@ -5,8 +5,7 @@ from decimal import Decimal
 DISCORD_SUPER_USER_IDS = [402536023483088896, 471510026696261632]
 
 # AI MODELS
-# OPENROUTER_MODELS = ["anthropic/claude-3.5-sonnet", "openai/chatgpt-4o-latest"]  # This doesn't work
-# OPENROUTER_ROUTE = 'fallback'
+DEFAULT_OPENROUTER_MODEL = 'anthropic/claude-3.5-sonnet:beta'
 DEFAULT_OPEN_AI_MODEL = 'chatgpt-4o-latest'
 DEFAULT_ANTHROPIC_MODEL = 'claude-3-5-sonnet-20241022'
 
@@ -28,8 +27,10 @@ MAX_HISTORY = 15  # TODO: rename this to something more descriptive
 TASKS_TO_GENERATE = 3
 
 # Context generation limits
+MAX_PENDING_PROPOSALS_IN_CONTEXT = 6
 MAX_ACCEPTANCES_IN_CONTEXT = 6
 MAX_REFUSALS_IN_CONTEXT = 6
+MAX_VERIFICATIONS_IN_CONTEXT = 6
 MAX_REWARDS_IN_CONTEXT = 10
 MAX_CHUNK_MESSAGES_IN_CONTEXT = 20
