+++ nodetools/utilities/generic_pft_utilities.py
@@ -28,8 +28,8 @@ from nodetools.utilities.transaction_requirements import TransactionRequirementS
 from nodetools.ai.openai import OpenAIRequestTool
 from nodetools.utilities.db_manager import DBConnectionManager
 from nodetools.utilities.credentials import CredentialManager
-import nodetools.utilities.constants as constants
-import nodetools.utilities.configuration as config
+import nodetools.configuration.constants as constants
+import nodetools.configuration.configuration as config
 from nodetools.utilities.base import BaseUtilities
 from nodetools.utilities.exceptions import *
 from loguru import logger
@@ -1291,173 +1291,6 @@ class GenericPFTUtilities(BaseUtilities):
 
         return core_task_df
 
-    # TODO: This is task management logic, move to task_management.py
-    def get_proposal_response_pairs(self, account_memo_detail_df):
-        """Convert account info into a DataFrame of proposed tasks and their responses (acceptances/refusals)
-        Args:
-            account_memo_detail_df: DataFrame containing account memo details
-            
-        Returns:
-            DataFrame with columns:
-                - proposal: The proposed task text
-                - response: The most recent response text (acceptance or refusal), or empty string if no response
-        """
-        task_frame = self.convert_all_account_info_into_simplified_task_frame(
-            account_memo_detail_df=account_memo_detail_df.sort_values('datetime')
-        )
-
-        if task_frame.empty:
-            return pd.DataFrame()
-
-        # Rename columns for clarity
-        task_frame.rename(columns={
-            'MemoType': 'task_id',
-            'MemoData': 'full_output',
-            'MemoFormat': 'user_account'
-        }, inplace=True)
-
-        # Get proposals and responses
-        proposals = task_frame[task_frame['task_type']==constants.TaskType.PROPOSAL.name].groupby('task_id').first()['full_output']
-        responses = task_frame[
-            (task_frame['task_type']==constants.TaskType.ACCEPTANCE.name) |
-            (task_frame['task_type']==constants.TaskType.REFUSAL.name)
-        ].groupby('task_id').last()['full_output']
-
-        # Combine proposals and responses, keeping all proposals
-        task_pairs = pd.DataFrame({'PROPOSAL': proposals})
-        task_pairs['RESPONSES'] = responses.fillna('')
-
-        task_pairs['RESPONSES'] = task_pairs['RESPONSES'].fillna('')
-
-        return task_pairs
-    
-    # TODO: This is task management logic, move to task_management.py
-    def get_proposal_acceptance_pairs(
-        self, 
-        account_memo_detail_df, 
-        include_pending=False, 
-        include_submitted_for_verification=True, 
-        include_rewarded=False
-    ):
-        """Convert account info into a DataFrame of proposed and accepted tasks.
-    
-        Args:
-            account_memo_detail_df: DataFrame containing account memo details
-            include_pending: If True, includes proposals without responses
-            include_submitted_for_verification: If True, includes tasks that have been submitted for verification
-            include_rewarded: If True, includes tasks that have been rewarded
-            
-        Returns:
-            DataFrame with two columns:
-                - proposal: The proposed task text (with 'PROPOSED PF ___' prefix removed)
-                - acceptance: The acceptance text (with 'ACCEPTANCE REASON ___' prefix removed)
-        """
-        # Get the base pairs from get_proposal_response_pairs
-        task_pairs = self.get_proposal_response_pairs(account_memo_detail_df)
-
-        if task_pairs.empty:
-            return pd.DataFrame()
-
-        if not include_rewarded:
-            # Get task IDs that have been rewarded
-            rewarded_tasks = account_memo_detail_df[
-                account_memo_detail_df['memo_data'].str.contains(constants.TaskType.REWARD.value, na=False)
-            ]['memo_type'].unique()
-
-            # Filter out rewarded tasks
-            task_pairs = task_pairs[~task_pairs.index.isin(rewarded_tasks)]
-
-        if not include_submitted_for_verification:
-            # Get task IDs that have been submitted for verification
-            verification_tasls = account_memo_detail_df[
-                account_memo_detail_df['memo_data'].str.contains(constants.TaskType.VERIFICATION_PROMPT.value, na=False)
-            ]['memo_type'].unique()
-
-            # Filter out verification tasks
-            task_pairs = task_pairs[~task_pairs.index.isin(verification_tasls)]
-
-        if include_pending:
-            # Keep acceptances and proposals without responses
-            acceptance_pairs = task_pairs[
-                (task_pairs['RESPONSES'].str.contains(constants.TaskType.ACCEPTANCE.value, na=False)) |
-                (task_pairs['RESPONSES'] == '')
-            ].copy()
-        else:
-            # Keep only acceptances
-            acceptance_pairs = task_pairs[
-                task_pairs['RESPONSES'].str.contains(constants.TaskType.ACCEPTANCE.value, na=False)
-            ].copy()
-
-        # Rename columns to match expected output
-        acceptance_pairs.rename(columns={
-            'PROPOSAL': 'proposal',
-            'RESPONSES': 'acceptance'
-        }, inplace=True)
-
-        # Clean up the text content
-        acceptance_pairs['acceptance'] = acceptance_pairs['acceptance'].apply(
-            lambda x: str(x).replace(constants.TaskType.ACCEPTANCE.value, '').replace('nan', '')
-        )
-        acceptance_pairs['proposal'] = acceptance_pairs['proposal'].apply(
-            lambda x: str(x).replace(constants.TaskType.PROPOSAL.value, '').replace('nan', '')
-        )
-
-        return acceptance_pairs
-
-    # TODO: This is task management logic, move to task_management.py
-    def get_proposal_refusal_pairs(self, account_memo_detail_df, exclude_refused=False):
-        """Get pairs of proposals and their refusals.
-        
-        Args:
-            all_account_info: DataFrame containing account memo details
-            exclude_refused: If True, excludes tasks that have any refusal response,
-                regardless of whether they were previously accepted
-            
-        Returns:
-            DataFrame with columns:
-                - refusal: The refusal text
-                - proposal: The corresponding proposal text
-            Indexed by memo_type
-        """
-        # Get base pairs from get_proposal_response_pairs
-        task_pairs = self.get_proposal_response_pairs(account_memo_detail_df)
-
-        if task_pairs.empty:
-            return pd.DataFrame()
-        
-        # Get tasks that have been rewarded and always exclude them
-        rewarded_tasks = account_memo_detail_df[
-            account_memo_detail_df['memo_data'].str.contains(constants.TaskType.REWARD.value, na=False)
-        ]['memo_type'].unique()
-
-        task_pairs = task_pairs[~task_pairs.index.isin(rewarded_tasks)]
-
-        if exclude_refused:
-            # Keep only proposals that have never been refused
-            # This checks the entire RESPONSES string for any refusal,
-            refusal_pairs = task_pairs[
-                ~task_pairs['RESPONSES'].str.contains(constants.TaskType.REFUSAL.value, na=False)
-            ].copy()
-        else:
-            # Keep all refusals
-            refusal_pairs = task_pairs.copy()
-
-        # Rename columns to match expected output
-        refusal_pairs.rename(columns={
-            'PROPOSAL': 'proposal',
-            'RESPONSES': 'refusal'
-        }, inplace=True)
-
-        # Clean up the text content
-        refusal_pairs['refusal'] = refusal_pairs['refusal'].apply(
-            lambda x: str(x).replace(constants.TaskType.REFUSAL.value, '').replace('nan', '')
-        )
-        refusal_pairs['proposal'] = refusal_pairs['proposal'].apply(
-            lambda x: str(x).replace(constants.TaskType.PROPOSAL.value, '').replace('nan', '')
-        )
-
-        return refusal_pairs
-
     def establish_post_fiat_tx_cache_as_hash_unique(self):
         dbconnx = self.db_connection_manager.spawn_sqlalchemy_db_connection_for_user(username=self.node_name)
         
@@ -1721,145 +1554,6 @@ class GenericPFTUtilities(BaseUtilities):
         #live_memo_tx['directional_pft'] = live_memo_tx['pft_sign']*live_memo_tx['pft_absolute_amount']
         live_memo_tx['simple_date']=pd.to_datetime(live_memo_tx['datetime'].apply(lambda x: x.strftime('%Y-%m-%d')))
         return live_memo_tx
-
-
-    def format_outstanding_tasks(self, outstanding_task_df):
-        """
-        Convert outstanding_task_df to a more legible string format for AI tools.
-        
-        Args:
-            outstanding_task_df: DataFrame containing outstanding tasks
-            
-        Returns:
-            Formatted string representation of the tasks
-        """
-        formatted_tasks = []
-        for idx, row in outstanding_task_df.iterrows():
-            task_str = f"Task ID: {idx}\n"
-            task_str += f"Proposal: {row['proposal']}\n"
-            task_str += f"Acceptance: {row['acceptance']}\n"
-            task_str += "-" * 50  # Separator
-            formatted_tasks.append(task_str)
-        
-        formatted_task_string =  "\n".join(formatted_tasks)
-        output_string="""OUTSTANDING TASKS
-    """+formatted_task_string
-        return output_string
-
-    def _calculate_weekly_reward_totals(self, specific_rewards):
-        """Calculate weekly reward totals with proper date handling.
-        
-        Returns DataFrame with weekly_total column indexed by date"""
-        # Calculate daily totals
-        daily_totals = specific_rewards[['directional_pft', 'simple_date']].groupby('simple_date').sum()
-
-        if daily_totals.empty:
-            logger.warning("No rewards data available to calculate weekly totals.")
-            return pd.DataFrame(columns=['weekly_total'])
-
-        # Extend date range to today
-        today = pd.Timestamp.today().normalize()
-        start_date = daily_totals.index.min()
-
-        if pd.isna(start_date):
-            logger.warning("Start date is NaT, cannot calculate weekly totals.")
-            return pd.DataFrame(columns=['weekly_total'])
-        
-        date_range = pd.date_range(
-            start=start_date,
-            end=today,
-            freq='D'
-        )
-
-        # Fill missing dates and calculate weekly totals
-        extended_daily_totals = daily_totals.reindex(date_range, fill_value=0)
-        extended_daily_totals = extended_daily_totals.resample('D').last().fillna(0)
-        extended_daily_totals['weekly_total'] = extended_daily_totals.rolling(7).sum()
-
-        # Return weekly totals
-        weekly_totals = extended_daily_totals.resample('W').last()[['weekly_total']]
-        weekly_totals.index.name = 'date'
-
-        # if weekly totals are NaN, set them to 0
-        weekly_totals = weekly_totals.fillna(0)
-
-        return weekly_totals
-    
-    def _pair_rewards_with_tasks(self, specific_rewards, all_account_info):
-        """Pair rewards with their original requests and proposals.
-        
-        Returns DataFrame with columns: memo_data, directional_pft, datetime, memo_type, request, proposal
-        """
-        # Get reward details
-        reward_details = specific_rewards[
-            ['memo_data', 'directional_pft', 'datetime', 'memo_type']
-        ].sort_values('datetime')
-
-        # Get original requests and proposals
-        task_requests = all_account_info[
-            all_account_info['memo_data'].apply(lambda x: constants.TaskType.REQUEST_POST_FIAT.value in x)
-        ].groupby('memo_type').first()['memo_data']
-
-        proposal_patterns = constants.TASK_PATTERNS[constants.TaskType.PROPOSAL]
-        task_proposals = all_account_info[
-            all_account_info['memo_data'].apply(lambda x: any(pattern in str(x) for pattern in proposal_patterns))
-        ].groupby('memo_type').first()['memo_data']
-
-        # Map requests and proposals to rewards
-        reward_details['request'] = reward_details['memo_type'].map(task_requests).fillna('No Request String')
-        reward_details['proposal'] = reward_details['memo_type'].map(task_proposals)
-
-        return reward_details
-
-    def get_reward_data(self, all_account_info):
-        """Get reward time series and task completion history.
-        
-        Args:
-            all_account_info: DataFrame containing account memo details
-            
-        Returns:
-            dict with keys:
-                - reward_ts: DataFrame of weekly reward totals
-                - reward_summaries: DataFrame containing rewards paired with original requests/proposals
-        """
-        # Get basic reward data
-        reward_responses = all_account_info[all_account_info['directional_pft'] > 0]
-        specific_rewards = reward_responses[
-            reward_responses.memo_data.apply(lambda x: "REWARD RESPONSE" in x)
-        ]
-
-        # Get weekly totals
-        weekly_totals = self._calculate_weekly_reward_totals(specific_rewards)
-
-        # Get reward summaries with context
-        reward_summaries = self._pair_rewards_with_tasks(
-            specific_rewards=specific_rewards,
-            all_account_info=all_account_info
-        )
-
-        return {
-            'reward_ts': weekly_totals,
-            'reward_summaries': reward_summaries
-        }
-
-    def format_reward_summary(self, reward_summary_df):
-        """
-        Convert reward summary dataframe into a human-readable string.
-        :param reward_summary_df: DataFrame containing reward summary information
-        :return: Formatted string representation of the rewards
-        """
-        formatted_rewards = []
-        for _, row in reward_summary_df.iterrows():
-            reward_str = f"Date: {row['datetime']}\n"
-            reward_str += f"Request: {row['request']}\n"
-            reward_str += f"Proposal: {row['proposal']}\n"
-            reward_str += f"Reward: {row['directional_pft']} PFT\n"
-            reward_str += f"Response: {row['memo_data'].replace(constants.TaskType.REWARD.value, '')}\n"
-            reward_str += "-" * 50  # Separator
-            formatted_rewards.append(reward_str)
-        
-        output_string = "REWARD SUMMARY\n\n" + "\n".join(formatted_rewards)
-        return output_string
 
     def get_latest_outgoing_context_doc_link(
             self, 
@@ -2168,128 +1862,6 @@ class GenericPFTUtilities(BaseUtilities):
             logger.error(f"GenericPFTUtilities.get_recent_user_memos: Failed to get recent user memos for account {account_address}: {e}")
             return json.dumps({})
 
-    def get_full_user_context_string(self, account_address: str, memo_history: pd.DataFrame):
-        """Get all core elements of a user's post fiat interactions.
-        Returns a context string even if some elements fail to generate.
-        Logs specific failures for debugging while allowing the function to continue.
-        """
-        MAX_ACCEPTANCES_IN_CONTEXT = constants.MAX_ACCEPTANCES_IN_CONTEXT
-        MAX_REFUSALS_IN_CONTEXT = constants.MAX_REFUSALS_IN_CONTEXT
-        MAX_REWARDS_IN_CONTEXT = constants.MAX_REWARDS_IN_CONTEXT
-        MAX_CHUNK_MESSAGES_IN_CONTEXT = constants.MAX_CHUNK_MESSAGES_IN_CONTEXT
-
-        memo_history.sort_values('datetime', inplace=True)
-
-        # Initialize core elements
-        core_element_outstanding_tasks = ''
-        core_element__refusal_frame = ''
-        core_element__last_10_rewards = ''
-        core_element_post_fiat_weekly_gen = ''
-        core_element__google_doc_text = ''
-        core_element__chunk_messages = ''
-
-        try:
-            # Retrieve outstanding task frame
-            proposal_acceptance_pairs_df = self.get_proposal_acceptance_pairs(account_memo_detail_df=memo_history, include_pending=True).tail(MAX_ACCEPTANCES_IN_CONTEXT)
-            if proposal_acceptance_pairs_df.empty:
-                logger.debug(f'GenericPFTUtilities.get_full_user_context_string: No proposals or acceptances found for {account_address}')
-                core_element_outstanding_tasks = "No proposals or acceptances found."
-            else:
-                core_element_outstanding_tasks = self.format_outstanding_tasks(outstanding_task_df=proposal_acceptance_pairs_df)
-
-        except Exception as e:
-            logger.error(f'GenericPFTUtilities.get_full_user_context_string: Exception for {account_address} while retrieving outstanding tasks: {e}')
-
-        try:
-            # Retrieve refusal frame
-            proposal_refusal_pairs_df = self.get_proposal_refusal_pairs(account_memo_detail_df=memo_history).tail(MAX_REFUSALS_IN_CONTEXT)
-            if proposal_refusal_pairs_df.empty:
-                logger.debug(f'GenericPFTUtilities.get_full_user_context_string: No proposals or refusals found for {account_address}')
-                core_element__refusal_frame = "No proposals or refusals found."
-            else:
-                core_element__refusal_frame = self.format_refusal_frame(refusal_frame_constructor=proposal_refusal_pairs_df)
-
-        except Exception as e:
-            logger.error(f'GenericPFTUtilities.get_full_user_context_string: Exception for {account_address} while retrieving refusal frame: {e}')
-
-        try:
-            # Retrieve rewards
-            reward_map = self.get_reward_data(all_account_info=memo_history)
-            weekly_totals, reward_summaries = reward_map['reward_ts'], reward_map['reward_summaries']
-
-            if reward_summaries.empty:
-                logger.debug(f'GenericPFTUtilities.get_full_user_context_string: No rewards found for {account_address}')
-                core_element__last_10_rewards = "No rewards found."
-            else:
-                core_element__last_10_rewards = self.format_reward_summary(reward_summaries.tail(MAX_REWARDS_IN_CONTEXT))
-            
-            if weekly_totals.empty:
-                logger.debug(f'GenericPFTUtilities.get_full_user_context_string: No weekly totals found for {account_address}')
-                core_element_post_fiat_weekly_gen = "No weekly totals found."
-            else:
-                core_element_post_fiat_weekly_gen = weekly_totals['weekly_total'].to_string()
-
-        except Exception as e:
-            logger.error(f'GenericPFTUtilities.get_full_user_context_string: Exception for {account_address} while retrieving rewards: {e}')
-
-        try:
-            # Retrieve latest google doc text
-            google_url = self.get_latest_outgoing_context_doc_link(account_address=account_address, memo_history=memo_history)
-            logger.debug(f"GenericPFTUtilities.get_full_user_context_string: Retrieved google doc link: {google_url}")
-            core_element__google_doc_text= self.get_google_doc_text(google_url)
-
-            logger.debug(f"GenericPFTUtilities.get_full_user_context_string: Retrieved google doc text: {core_element__google_doc_text}")
-
-        except Exception as e:
-            logger.error(f'GenericPFTUtilities.get_full_user_context_string: Exception for {account_address} while retrieving google doc text: {e}')
-
-        try:
-            # Retrieve chunk messages
-            core_element__chunk_messages = self.get_recent_user_memos(
-                account_address=account_address, 
-                num_messages=MAX_CHUNK_MESSAGES_IN_CONTEXT
-            )
-
-        except Exception as e:
-            logger.error(f'GenericPFTUtilities.get_full_user_context_string: Exception for {account_address} while retrieving chunk messages: {e}')
-
-        # Format final context string
-        current_date = datetime.datetime.now().strftime('%Y-%m-%d')
-        final_context_string = f"""The current date is {current_date}
-        
-        USERS CORE OUTSTANDING TASKS ARE AS FOLLOWS:
-        <OUTSTANDING TASKS START HERE>
-        {core_element_outstanding_tasks}
-        <OUTSTANDING TASKS END HERE>
-
-        THESE ARE TASKS USER HAS RECENTLY REFUSED ALONG WITH REASONS
-        <REFUSED TASKS START HERE>
-        {core_element__refusal_frame}
-        <REFUSED TASKS END HERE>
-        
-        THESE ARE TASKS USER HAS RECENTLY COMPLETED ALONG WITH REWARDS
-        <REWARDED TASKS START HERE>
-        {core_element__last_10_rewards}
-        <REWARDED TASKS END HERE>
-        
-        HERE IS THE USERS RECENT POST FIAT OUTPUT SUMMED AS A WEEKLY TIMESERIES
-        <POST FIAT GENERATION STARTS HERE>
-        {core_element_post_fiat_weekly_gen}
-        <POST FIAT GENERATION ENDS HERE>
-        
-        THE FOLLOWING IS THE PRIMARY CONTENT OF THE USERS CONTEXT DOCUMENT AND PLANNING
-        <USER CONTEXT AND PLANNING STARTS HERE>
-        {core_element__google_doc_text}
-        <USER CONTEXT AND PLANNING ENDS HERE>
-        
-        THE FOLLOWING ARE THE RECENT LONG FORM DIALOGUES WITH THE USER
-        <USER LONG FORM DIALOGUE>
-        {core_element__chunk_messages}
-        <USER LONG FORM DIALOGUE ENDS>
-        """
-
-        return final_context_string
-
     def create_xrp_wallet(self):
         test_wallet = Wallet.create()
         classic_address= test_wallet.classic_address
@@ -2301,64 +1873,6 @@ STORE YOUR WALLET SECRET IN AN OFFLINE PREFERABLY NON DIGITAL LOCATION
 THIS MESSAGE WILL AUTO DELETE IN 60 SECONDS
 """
         return output_string
-
-#     # TODO: Move to discord module
-#     def generate_basic_balance_info_string_for_account_address(self, account_address):
-#         try:
-#             memo_history =self.get_account_memo_history(account_address=account_address)
-#         except:
-#             pass
-#         monthly_pft_reward_avg=0
-#         weekly_pft_reward_avg=0
-#         try:
-#             reward_ts = self.get_reward_data(all_account_info=memo_history)
-#             monthly_pft_reward_avg = list(reward_ts['reward_ts'].tail(4).mean())[0]
-#             weekly_pft_reward_avg = list(reward_ts['reward_ts'].tail(1).mean())[0]
-#         except:
-#             pass
-#         number_of_transactions =0
-#         try:
-#             number_of_transactions = len(memo_history['memo_type'])
-#         except:
-#             pass
-#         user_name=''
-#         try:
-#             user_name = list(memo_history[memo_history['direction']=='OUTGOING']['memo_format'].mode())[0]
-#         except:
-#             pass
-        
-#         client = JsonRpcClient(self.primary_endpoint)
-        
-#         # Get XRP balance
-#         acct_info = AccountInfo(
-#             account=account_address,
-#             ledger_index="validated"
-#         )
-#         response = client.request(acct_info)
-#         xrp_balance=0
-#         try:
-#             xrp_balance = int(response.result['account_data']['Balance'])/1_000_000
-#         except:
-#             pass
-#         pft_balance= 0 
-#         try:
-#             account_lines = AccountLines(
-#                 account=account_address,
-#                 ledger_index="validated"
-#             )
-#             account_line_response = client.request(account_lines)
-#             pft_balance = [i for i in account_line_response.result['lines'] if i['account']=='rnQUEEg8yyjrwk9FhyXpKavHyCRJM9BDMW'][0]['balance']
-#         except:
-#             pass
-#         account_info_string =f"""ACCOUNT INFO for  {account_address}
-# LIKELY ALIAS:     {user_name}
-# XRP BALANCE:      {xrp_balance}
-# PFT BALANCE:      {pft_balance}
-# NUM PFT MEMO TX: {number_of_transactions}
-# PFT MONTHLY AVG:  {monthly_pft_reward_avg}
-# PFT WEEKLY AVG:   {weekly_pft_reward_avg}
-# """
-#         return account_info_string
     
     def get_pft_balance(self, address: str) -> float:
         """Get PFT balance for an account.
@@ -2427,89 +1941,6 @@ THIS MESSAGE WILL AUTO DELETE IN 60 SECONDS
         """
         balance = self.get_xrp_balance(address)
         return (balance >= minimum_xrp_balance, balance)
-
-    # TODO: Refactor using get_verification_df as reference (pftpyclient/task_manager/basic_tasks.py)
-    def get_verification_df(self, account_memo_detail_df):
-        """Takes the account memo dataframe and converts into outstanding verification tasks.
-        
-        Args:
-            account_memo_detail_df: DataFrame containing account memo details
-        Returns:
-            DataFrame with verification requirements
-        """
-        all_memos = account_memo_detail_df.copy()  # TODO: This copy might be unnecessary
-
-        # Get rewarded task IDs to exclude
-        rewarded_tasks = all_memos[
-            all_memos['memo_data'].apply(lambda x: constants.TaskType.REWARD.value in str(x))
-        ]['memo_type'].unique()
-
-        # Get most recent memos excluding rewarded tasks
-        most_recent_memos = (all_memos[~all_memos['memo_type'].isin(rewarded_tasks)]
-                             .sort_values('datetime')
-                             .groupby('memo_type')
-                             .last()
-                             .copy())
-
-        # Map task IDs to original proposals
-        proposal_patterns = constants.TASK_PATTERNS[constants.TaskType.PROPOSAL]
-        task_id_to_original_task_map = (all_memos[
-            all_memos['memo_data'].apply(lambda x: any(pattern in str(x) for pattern in proposal_patterns))
-        ][['memo_data','memo_type','memo_format']]
-            .groupby('memo_type')
-            .first()['memo_data'])
-
-        # Filter for verification prompts
-        verification_requirements = (most_recent_memos[
-            most_recent_memos['memo_data'].apply(lambda x: constants.TaskType.VERIFICATION_PROMPT.value in str(x))
-        ][['memo_data','memo_format']]
-            .reset_index()
-            .copy())
-
-        verification_requirements['original_task'] = verification_requirements['memo_type'].map(task_id_to_original_task_map)
-
-        return verification_requirements
-
-    def format_outstanding_verification_df(self, verification_requirements):
-        """
-        Format the verification_requirements dataframe into a string.
-
-        Args:
-        verification_requirements (pd.DataFrame): DataFrame containing columns 
-                                                'memo_type', 'memo_data', 'memo_format', and 'original_task'
-
-        Returns:
-        str: Formatted string of verification requirements
-        """
-        formatted_output = "VERIFICATION REQUIREMENTS\n"
-        for _, row in verification_requirements.iterrows():
-            formatted_output += f"Task ID: {row['memo_type']}\n"
-            formatted_output += f"Verification Prompt: {row['memo_data']}\n"
-            formatted_output += f"Original Task: {row['original_task']}\n"
-            formatted_output += "-" * 50 + "\n"
-        return formatted_output
-
-    def create_full_outstanding_pft_string(self, account_address):
-        """ 
-        This takes in an account address and outputs the current state of its outstanding tasks.
-        Returns empty string for accounts with no PFT-related transactions.
-        """ 
-        memo_history = self.get_account_memo_history(account_address=account_address, pft_only=True)
-        if memo_history.empty:
-            return ""
-        
-        memo_history.sort_values('datetime', inplace=True)
-        outstanding_task_df = self.get_proposal_acceptance_pairs(
-            account_memo_detail_df=memo_history, 
-            include_pending=True,
-            include_submitted_for_verification=False,
-            include_rewarded=False
-        )
-        task_string = self.format_outstanding_tasks(outstanding_task_df)
-        verification_df = self.get_verification_df(account_memo_detail_df=memo_history)
-        verification_string = self.format_outstanding_verification_df(verification_requirements=verification_df)
-        full_postfiat_outstanding_string=f"{task_string}\n{verification_string}"
-        return full_postfiat_outstanding_string
 
     def extract_transaction_info_from_response_object(self, response):
         """
@@ -2820,129 +2251,7 @@ THIS MESSAGE WILL AUTO DELETE IN 60 SECONDS
         """
         return re.sub(r'^chunk_\d+__', '', memo_data)
 
-    def format_tasks_for_discord(self, input_text: str):
-        """
-        Format task list for Discord with proper formatting and emoji indicators
-        Returns a list of formatted chunks ready for Discord sending
-        """
-        # Handle empty input
-        if not input_text or input_text.strip() == "OUTSTANDING TASKS":
-            return ["```ansi\n\u001b[1;33m=== OUTSTANDING TASKS ===\u001b[0m\n\u001b[0;37mNo outstanding tasks found.\u001b[0m\n```"]
-
-        # Split into main sections first
-        if "VERIFICATION REQUIREMENTS" in input_text:
-            tasks_section, verification_section = input_text.split("VERIFICATION REQUIREMENTS", 1)
-        else:
-            tasks_section, verification_section = input_text, ""
-
-        # Check if tasks section is empty (just the header)
-        tasks_section = tasks_section.strip()
-        if tasks_section == "OUTSTANDING TASKS":
-            return ["```ansi\n\u001b[1;33m=== OUTSTANDING TASKS ===\u001b[0m\n\u001b[0;37mNo outstanding tasks found.\u001b[0m\n```"]
-        
-        # Process tasks - remove the header line and split remaining tasks
-        tasks_lines = tasks_section.strip().split('\n', 1)[1]  # Skip the "OUTSTANDING TASKS" header
-        tasks = tasks_lines.split('--------------------------------------------------')
-        tasks = [t.strip() for t in tasks if t.strip()]  # Remove empty tasks and whitespace
-        
-        # Initialize formatted output parts
-        formatted_parts = []
-        current_chunk = ["```ansi\n\u001b[1;33m=== OUTSTANDING TASKS ===\u001b[0m\n"]
-        current_chunk_size = len(current_chunk[0])
-        
-        def add_to_chunks(content):
-            nonlocal current_chunk, current_chunk_size
-            content_size = len(content) + 1  # +1 for newline
-            
-            if current_chunk_size + content_size > 1900:
-                current_chunk.append("```")
-                formatted_parts.append("\n".join(current_chunk))
-                current_chunk = ["```ansi\n"]
-                current_chunk_size = len(current_chunk[0])
-                
-            current_chunk.append(content)
-            current_chunk_size += content_size
-        
-        # Process tasks
-        for task in tasks:
-            if not task.strip():
-                continue
-                
-            task_id_match = re.search(r'Task ID: ([0-9A-Za-z\-_:]+)', task)
-            proposal_match = re.search(r'Proposal: (.+?)(?=\nAcceptance:|$)', task, re.DOTALL)
-            acceptance_match = re.search(r'Acceptance: ?(.*?)(?=\n|$)', task, re.DOTALL)
-            priority_match = re.search(r'\.\. (\d+)', task)
-            
-            if not all([task_id_match, proposal_match, acceptance_match, priority_match]):
-                continue
-                
-            datetime_str = task_id_match.group(1).split('__')[0]
-            try:
-                date_obj = datetime.datetime.strptime(datetime_str, '%Y-%m-%d_%H:%M')
-                formatted_date = date_obj.strftime('%d %b %Y %H:%M')
-            except ValueError:
-                formatted_date = datetime_str
-
-            # Format acceptance status
-            acceptance_text = acceptance_match.group(1).strip()
-            acceptance_display = acceptance_text if acceptance_text else "(Pending)"
-            
-            # Format task components
-            task_parts = [
-                f"\u001b[1;36m📌 Task {task_id_match.group(1)}\u001b[0m",
-                f"\u001b[0;37mDate: {formatted_date}\u001b[0m",
-                f"\u001b[0;32mPriority: {priority_match.group(1)}\u001b[0m",
-                f"\u001b[1;37mProposal:\u001b[0m\n{proposal_match.group(1).strip()}",
-                f"\u001b[1;37mAcceptance:\u001b[0m\n{acceptance_display}",
-                "─" * 50
-            ]
-            
-            # Add each part to chunks
-            for part in task_parts:
-                add_to_chunks(part)
-
-        # Process verification section if it exists
-        if verification_section:
-            add_to_chunks("\n")  # Add spacing
-            add_to_chunks(f"\u001b[1;33m=== VERIFICATION REQUIREMENTS ===\u001b[0m")
-            
-            # Process each verification requirement
-            v_tasks = verification_section.split('--------------------------------------------------')
-            for vtask in v_tasks:
-                if not vtask.strip():
-                    continue
-                    
-                v_task_id_match = re.search(r'Task ID: ([0-9A-Za-z\-_:]+)', vtask)
-                v_prompt_match = re.search(r'Verification Prompt: (.+?)(?=\nOriginal Task:|$)', vtask, re.DOTALL)
-                v_original_match = re.search(r'Original Task: (.+?)(?=\n|$)', vtask, re.DOTALL)
-                
-                if not all([v_task_id_match, v_prompt_match]):
-                    continue
-                    
-                datetime_str = v_task_id_match.group(1).split('__')[0]
-                try:
-                    date_obj = datetime.datetime.strptime(datetime_str, '%Y-%m-%d_%H:%M')
-                    formatted_date = date_obj.strftime('%d %b %Y %H:%M')
-                except ValueError:
-                    formatted_date = datetime_str
-                
-                v_parts = [
-                    f"\u001b[1;36mTask {v_task_id_match.group(1)}\u001b[0m",
-                    f"\u001b[0;37mDate: {formatted_date}\u001b[0m",
-                    f"\u001b[1;37mPrompt:\u001b[0m\n{v_original_match.group(1).strip().replace(constants.TaskType.PROPOSAL.value, '')}",
-                    f"\u001b[1;37mVerification Prompt:\u001b[0m\n{v_prompt_match.group(1).strip().replace(constants.TaskType.VERIFICATION_PROMPT.value, '')}",
-                    "─" * 50
-                ]
-                
-                for part in v_parts:
-                    add_to_chunks(part)
-        
-        # Finalize last chunk
-        current_chunk.append("```")
-        formatted_parts.append("\n".join(current_chunk))
-        
-        return formatted_parts
-
+    # TODO: Refactor, add documentation and move to a different module
     def output_postfiat_foundation_node_leaderboard_df(self):
         """ This generates the full Post Fiat Foundation Leaderboard """ 
         all_accounts = self.get_all_account_pft_memo_data()
@@ -3100,6 +2409,7 @@ THIS MESSAGE WILL AUTO DELETE IN 60 SECONDS
         final_leaderboard.index.name = 'Foundation Node Leaderboard as of '+datetime.datetime.now().strftime('%Y-%m-%d')
         return final_leaderboard
 
+    # TODO: Refactor, add documentation and move to a different module
     def format_and_write_leaderboard(self):
         """ This loads the current leaderboard df and writes it""" 
         def format_leaderboard_df(df):
